Increase Axis Research Sharing bonus from .05 to .1
This has it match Vanilla values and was the only one not at 10% for some reason. Also added in the ones from the new DLC

Note: Vichy's buff is inconsistent between what we give them and Vanilla. Vanilla gives them a custom one for Aircraft but we add them to the axis