New Italy Joining War
- New Game Rule option, set as the default for Italy to join the war.
- Instead of a set days after going to war with France, its instead 45 days after EITHER Around the Maginot OR France loses 1 Province.

The intention of this change is to allow for a longer Phony War/Invasion of Norway without forcing Germany directly into the Benelux, while also making it so if Germany stalls out in their Blitzkrieg that Italy will be in trouble. Starting with 45 days, may change it to 30/60 later. This should allow people who are bad at Germany (Swimmy) to not ruin their teams chance of winning when they don't move from Poland to Belgium quick enough